+++ backend/src/routes/auth.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { login, register } from "../controllers/authController";
+import { login, register } from "../controllers/auth";
 
 const router = express.Router();
